backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "boto3==1.40.55",
    "botocore==1.40.55",
    "celery==5.4.0",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi==0.119.0",
    "fastapi-cli==0.0.13",
    "fastapi-cloud-cli==0.3.1",
    "flower==2.0.1",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "jmespath==1.0.1",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "psycopg2-binary==2.9.11",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-settings==2.1.0",
    "pygments==2.19.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.3",
    "redis==5.2.1",
    "requests==2.32.5",
    "rich==14.2.0",
    "rich-toolkit==0.15.1",
    "rignore==0.7.1",
    "s3transfer==0.14.0",
    "sentry-sdk==2.42.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.44",
    "starlette==0.48.0",
    "tornado==6.4.2",
    "typer==0.19.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
]
